refactor(modal): separate UI and Logic by `useModal` hook
