+++ packages/components/modal/src/index.ts
@@ -34,3 +34,5 @@ export { ModalHeader } from "./modal-header"
 export type { ModalHeaderProps } from "./modal-header"
 export { ModalOverlay } from "./modal-overlay"
 export type { ModalOverlayProps } from "./modal-overlay"
+export { useModal } from "./use-modal"
+export type { UseModalProps, UseModalReturn } from "./use-modal"
